requirements.txt
PyQt5>=5.15
mido>=1.3.0
python-rtmidi>=1.5.0
numpy>=1.21.0
sounddevice>=0.4.0
